Refactor `images` package to `cache`
