+++ admin/app/views/admin.scala.html
@@ -30,7 +30,7 @@
                         <li><a href="@controllers.admin.routes.RadiatorController.renderRadiator()">Radiator</a></li>
                         <li><a href="@controllers.admin.routes.RadiatorController.renderRadiator()?features=external">Radiator (external)</a></li>
                         <li><a href="@controllers.admin.routes.RedirectController.redirect()">Redirects</a></li>
-                        <li><a href="@controllers.images.routes.ImageDecacheController.renderImageDecacheForm()">Clear images</a></li>
+                        <li><a href="@controllers.cache.routes.ImageDecacheController.renderImageDecacheForm()">Clear images</a></li>
                         <li><a href="@controllers.admin.routes.CssReportController.entry()">CSS Selector Usage</a></li>
                     </ul>
                 </div>
